made the mask longs as pseudo-constants, and thinking about how to do custom hash and dictionaries

+++ TessellationAndVoxelizationGeometryLibrary/Voxelization/Voxel.cs
@@ -164,7 +164,7 @@ namespace TVGL.Voxelization
         }
 
 
-        internal IEnumerable<double[]> GetVoxels(byte targetFlags, VoxelizedSolid voxelizedSolid, int level)
+        internal IEnumerable<double[]> GetVoxels(long targetFlags, VoxelizedSolid voxelizedSolid, int level)
         {
             foreach (var voxel in Voxels)
             {
